+++ Core/Crash/chfcts.hpp
@@ -20,7 +20,7 @@ struct CRData
     CPU::x64::CR4 cr4;
     CPU::x64::CR8 cr8;
     CPU::x64::EFER efer;
-    uint64_t dr0, dr1, dr2, dr3, dr6;
+    uintptr_t dr0, dr1, dr2, dr3, dr6;
     CPU::x64::DR7 dr7;
 
     long ID;
@@ -42,7 +42,7 @@ struct CRData
     CPU::x32::CR4 cr4;
     CPU::x32::CR8 cr8;
     CPU::x32::EFER efer;
-    uint64_t dr0, dr1, dr2, dr3, dr6;
+    uintptr_t dr0, dr1, dr2, dr3, dr6;
     CPU::x32::DR7 dr7;
 
     long ID;
